+++ src/hooks/index.js
@@ -1,1 +1,1 @@
-export {};
+export { useForm } from 'hooks/useForm';
