+++ src/pages/SignUp/SignUp.jsx
@@ -1,34 +1,16 @@
-import { useState } from 'react';
 import { Form } from 'components';
-import { auth } from 'api';
+import { useForm } from 'hooks';
 import * as S from './style';
 
 export function SignUp() {
-  const [formData, setFormData] = useState({
-    fullName: '',
-    email: '',
-    password: '',
+  const { onChange, onSubmit } = useForm({
+    initialState: {
+      fullName: '',
+      email: '',
+      password: '',
+    },
+    authType: 'signup',
   });
-
-  const onChange = ({ target }) => {
-    const { name, value } = target;
-
-    setFormData({
-      ...formData,
-      [name]: value,
-    });
-  };
-
-  const onSubmit = async e => {
-    e.preventDefault();
-
-    try {
-      const response = await auth.signup({ ...formData });
-      console.log(response);
-    } catch (err) {
-      console.error(err);
-    }
-  };
 
   const info = {
     title: '회원가입',
